Kinematic rigidbody
- Disconnected Vehicle collider from Vehicle and added kinematic rigidbody
